% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunningStats.R
\name{calculateSlope}
\alias{calculateSlope}
\title{Calculating the slope profile}
\usage{
calculateSlope(tr)
}
\arguments{
\item{Track-class}{}
}
\value{
DataFrame
}
\description{
The function calculates the splope profile of a track by smoothing the elevation profile by applying a low-pass filter. The slope profile is determined by local minimas and maximas.
}
\examples{
\dontrun{
calculateSlope(tr)
}
}

