% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunningStats.R
\name{getStats}
\alias{getStats}
\title{Getting Track statistics}
\usage{
getStats(tr)
}
\arguments{
\item{Track-class}{}
}
\value{
DataFrame
}
\description{
The function returns track statistics including Date, Start-time, End-time, Distance, Minutes, min/km, SpeedMax[km/h] and weather data.
}
\examples{
\dontrun{
getStats(tr)
}
}

