% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunningStats.R
\name{getTr}
\alias{getTr}
\title{Loading GPX File}
\usage{
getTr(file)
}
\arguments{
\item{Path}{to GPX file}
}
\value{
Track-class
}
\description{
The function returns track statistics including Date, Start-time, End-time, Meters, Minutes, min/km, SpeedMax[km/h] and weather data.
}
\examples{
\dontrun{
getTr(file)
}
}

