% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openweathermap.R
\name{getWeatherInfo}
\alias{getWeatherInfo}
\title{Get Weather Information}
\usage{
getWeatherInfo(track)
}
\arguments{
\item{Track-class}{}
}
\value{
DataFrame
}
\description{
The function queries weather data for the Track from OpenWeatherMap.org
}
\examples{
\dontrun{
getWeatherInfo(track)
}
}

