% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunningStats.R
\name{plotEleSpeed}
\alias{plotEleSpeed}
\title{Plotting Elevation against Speed}
\usage{
plotEleSpeed(tr)
}
\arguments{
\item{Track-class}{}
}
\value{
plot
}
\description{
The function plots the elevation profile and speed values.
}
\examples{
\dontrun{
plotEleSpeed(track=track)
}
}

