% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunningStats.R
\name{plotMap}
\alias{plotMap}
\title{Plot Map}
\usage{
plotMap(tr, x, y)
}
\arguments{
\item{Track-class,}{x (lat), y (lon)}
}
\value{
Plot
}
\description{
The function returns track statistics including Date, Start-time, End-time, Distance, Minutes, min/km, SpeedMax[km/h] and weather data.
}
\examples{
\dontrun{
plotMap(tr, x, y)
}
}

