% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speed.R
\name{plotSpe}
\alias{plotSpe}
\title{Plot Speed}
\usage{
plotSpe(tr)
}
\arguments{
\item{Track-class}{}
}
\value{
plot
}
\description{
The function plots the speed data on a map by applying a scale colour gradient
}
\examples{
\dontrun{
getStats(track)
}
}

